# Remove outdated `vformat` comments/workarounds in `time.cpp`