`Time` methods returning Strings with both date and time values included were calling `vformat` twice, as older `vformat` implementation limited how many arguments could be passed in.  Updated these methods to just call `vformat` once.

This also makes these methods around 10-20% faster, compared with below gdscript:

```gdscript
func _ready() -> void:
	var start := Time.get_ticks_msec()
	for i in 100000:
		Time.get_datetime_string_from_system()
	var end := Time.get_ticks_msec()
	print("get_datetime_string_from_system: %dms" % [end - start])

	var unix_time = Time.get_unix_time_from_system()
	start = Time.get_ticks_msec()
	for i in 100000:
		Time.get_datetime_string_from_unix_time(unix_time)
	end = Time.get_ticks_msec()
	print("get_datetime_string_from_unix_time: %dms" % [end - start])
	
	var dictionary := {
		"year": 2025,
		"month": 1,
		"day": 2,
		"hour": 3,
		"minute": 4,
		"second": 5,
	}
	start = Time.get_ticks_msec()
	for i in 100000:
		Time.get_datetime_string_from_datetime_dict(dictionary, true)
	end = Time.get_ticks_msec()
	print("get_datetime_string_from_datetime_dict: %dms" % [end - start])
```

old:
```
get_datetime_string_from_system: 269ms
get_datetime_string_from_unix_time: 207ms
get_datetime_string_from_datetime_dict: 418ms
```

new:
```
get_datetime_string_from_system: 222ms
get_datetime_string_from_unix_time: 162ms
get_datetime_string_from_datetime_dict: 370ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->