@clayjohn 2025-03-10
> as older vformat implementation limited how many arguments could be passed in

Do you know what compiler/version had these limitations and when the limitation was lifted?