@aaronp64 2025-03-10
Looks like it only took 5 arguments until https://github.com/godotengine/godot/pull/65760